Remove referrers test and merge time stats tests
