+++ example/src/androidTest/java/org/wordpress/android/fluxc/release/ReleaseStack_AppComponent.java
@@ -31,7 +31,6 @@ public interface ReleaseStack_AppComponent {
     void inject(ReleaseStack_ActivityLogTestJetpack test);
     void inject(ReleaseStack_InsightsTestJetpack test);
     void inject(ReleaseStack_TimeStatsTestJetpack test);
-    void inject(ReleaseStack_StatsReferrersTestJetpack test);
     void inject(ReleaseStack_CommentTestWPCom test);
     void inject(ReleaseStack_CommentTestXMLRPC test);
     void inject(ReleaseStack_DiscoveryTest test);
